+++ packages/shared/src/types/index.ts
@@ -40,6 +40,10 @@ export type ConnectionConfig = {
    */
   awsRegion?: string;
   /**
+   * ARN of the role that the proxy-server should assume to sign requests.
+   */
+  awsAssumeRoleArn?: string;
+  /**
    * Number of milliseconds before aborting a request.
    * By default, undefined.
    */